Move tests for magics that are terminal-specific to their own file.
